backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
authlib==1.3.1
itsdangerous==2.2.0
python-dotenv==1.0.1
httpx==0.27.0
